+++ connect/src/main/java/com/stripe/android/connect/webview/serialization/PaymentsPropsJs.kt
@@ -1,7 +1,7 @@
 package com.stripe.android.connect.webview.serialization
 
 import com.stripe.android.connect.PaymentsProps
-import com.stripe.android.connect.PrivateBetaConnectSDK
+import com.stripe.android.connect.PreviewConnectSDK
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -44,7 +44,7 @@ internal data class PaymentsPropsJs(
     )
 }
 
-@OptIn(PrivateBetaConnectSDK::class)
+@OptIn(PreviewConnectSDK::class)
 internal fun PaymentsProps.toJs(): PaymentsPropsJs {
     return PaymentsPropsJs(
         setDefaultFilters = defaultFilters?.let { filters ->