+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## XX.XX.XX - 20XX-XX-XX
 
+### Connect
+* [Fixed] Renamed `PrivateBetaConnectSDK` to `PreviewConnectSDK`.
+
 ## 21.27.0 - 2025-09-22
 
 ### Connect